refactor and fixes

+++ BrightData.UnitTests/VectorisationTests.cs
@@ -27,7 +27,7 @@ namespace BrightData.UnitTests
             var indexList = _context.CreateIndexList(1, 2, 3);
             var weightedIndexList = _context.CreateWeightedIndexList((1, 0.1f), (2, 0.5f), (3, 1f));
             var vector = _context.CreateReadOnlyVector(3, 0.25f);
-            builder.AddRow(false, 1, 2, 3, 4, 5, 6, indexList, weightedIndexList, vector);
+            builder.AddRow(false, (sbyte)1, (decimal)2, (double)3, (float)4, 5, (long)6, indexList, weightedIndexList, vector);
             return builder.BuildInMemory();
         }
